+++ swagger_parser/test/e2e/tests/basic/excluded_tags/expected_files/models/get_api_v1_empty_tags_endpoint_response.dart
@@ -11,7 +11,7 @@ part 'get_api_v1_empty_tags_endpoint_response.g.dart';
 class GetApiV1EmptyTagsEndpointResponse
     with _$GetApiV1EmptyTagsEndpointResponse {
   const factory GetApiV1EmptyTagsEndpointResponse({
-    String? emptyTagsField,
+    @JsonKey(includeIfNull: false) String? emptyTagsField,
   }) = _GetApiV1EmptyTagsEndpointResponse;
 
   factory GetApiV1EmptyTagsEndpointResponse.fromJson(