+++ swagger_parser/test/e2e/tests/basic/excluded_tags_with_schemas/expected_files/models/category.dart
@@ -10,8 +10,8 @@ part 'category.g.dart';
 @Freezed()
 class Category with _$Category {
   const factory Category({
-    int? id,
-    String? name,
+    @JsonKey(includeIfNull: false) int? id,
+    @JsonKey(includeIfNull: false) String? name,
   }) = _Category;
 
   factory Category.fromJson(Map<String, Object?> json) =>