+++ swagger_parser/test/e2e/tests/corrector/expected_files/models/p1_class_dto.dart
@@ -10,7 +10,7 @@ part 'p1_class_dto.g.dart';
 @Freezed()
 class P1ClassDto with _$P1ClassDto {
   const factory P1ClassDto({
-    DateTime? test,
+    @JsonKey(includeIfNull: false) DateTime? test,
   }) = _P1ClassDto;
 
   factory P1ClassDto.fromJson(Map<String, Object?> json) =>