+++ swagger_parser/test/e2e/tests/request_unnamed_types/expected_files/models/get_test2_response.dart
@@ -11,8 +11,8 @@ part 'get_test2_response.g.dart';
 class GetTest2Response with _$GetTest2Response {
   const factory GetTest2Response({
     required List<String> list,
-    required String? name,
-    String? lastname,
+    @JsonKey(includeIfNull: true) required String? name,
+    @JsonKey(includeIfNull: false) String? lastname,
   }) = _GetTest2Response;
 
   factory GetTest2Response.fromJson(Map<String, Object?> json) =>